Enhancement: Better handle segmentation to bridge <---> GDB
Since the inception of BREAD, there has been a complicated issue to
handle: segmentation.

In real mode, addresses are always calculated using SEG:OFF, and this
was handed over as-is to GDB (which does not comprehend segmentation).

To work around GDB's limitations, whenever GDB requested memory
read/write operations, BREAD had to make educated guesses to
understand the address's origin and then convert it to physical form
using the appropriate segment. However, this approach made the code
convoluted and error-prone. This commit addresses that.

Instead of delivering all registers 'as-is,' registers that involve
segments (EIP, ESP, and EBP) are now returned to GDB already
converted to physical addresses using their respective segments (CS
and SS). This ensures that GDB always knows the correct physical
addresses and requests them directly, eliminating the need for future
conversions within the bridge.